Moves the watcher to its own file
